+++ nodejs/src/delivery/delete_entries.ts
@@ -1,23 +1,25 @@
 /* eslint-disable */
 import _m0 from "protobufjs/minimal";
+import { EntryFilter } from "./entry_filter";
 
 export const protobufPackage = "delivery";
 
-export interface DeleteEntryRequest {
+export interface DeleteEntriesRequest {
   type: string;
   tenantId: string;
   id: string;
+  filter: EntryFilter | undefined;
 }
 
-export interface DeleteEntryResponse {
+export interface DeleteEntriesResponse {
 }
 
-function createBaseDeleteEntryRequest(): DeleteEntryRequest {
-  return { type: "", tenantId: "", id: "" };
+function createBaseDeleteEntriesRequest(): DeleteEntriesRequest {
+  return { type: "", tenantId: "", id: "", filter: undefined };
 }
 
-export const DeleteEntryRequest = {
-  encode(message: DeleteEntryRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
+export const DeleteEntriesRequest = {
+  encode(message: DeleteEntriesRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
     if (message.type !== "") {
       writer.uint32(10).string(message.type);
     }
@@ -27,13 +29,16 @@ export const DeleteEntryRequest = {
     if (message.id !== "") {
       writer.uint32(26).string(message.id);
     }
+    if (message.filter !== undefined) {
+      EntryFilter.encode(message.filter, writer.uint32(34).fork()).ldelim();
+    }
     return writer;
   },
 
-  decode(input: _m0.Reader | Uint8Array, length?: number): DeleteEntryRequest {
+  decode(input: _m0.Reader | Uint8Array, length?: number): DeleteEntriesRequest {
     const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
     let end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseDeleteEntryRequest();
+    const message = createBaseDeleteEntriesRequest();
     while (reader.pos < end) {
       const tag = reader.uint32();
       switch (tag >>> 3) {
@@ -46,6 +51,9 @@ export const DeleteEntryRequest = {
         case 3:
           message.id = reader.string();
           break;
+        case 4:
+          message.filter = EntryFilter.decode(reader, reader.uint32());
+          break;
         default:
           reader.skipType(tag & 7);
           break;
@@ -54,48 +62,53 @@ export const DeleteEntryRequest = {
     return message;
   },
 
-  fromJSON(object: any): DeleteEntryRequest {
+  fromJSON(object: any): DeleteEntriesRequest {
     return {
       type: isSet(object.type) ? String(object.type) : "",
       tenantId: isSet(object.tenantId) ? String(object.tenantId) : "",
       id: isSet(object.id) ? String(object.id) : "",
+      filter: isSet(object.filter) ? EntryFilter.fromJSON(object.filter) : undefined,
     };
   },
 
-  toJSON(message: DeleteEntryRequest): unknown {
+  toJSON(message: DeleteEntriesRequest): unknown {
     const obj: any = {};
     message.type !== undefined && (obj.type = message.type);
     message.tenantId !== undefined && (obj.tenantId = message.tenantId);
     message.id !== undefined && (obj.id = message.id);
+    message.filter !== undefined && (obj.filter = message.filter ? EntryFilter.toJSON(message.filter) : undefined);
     return obj;
   },
 
-  create(base?: DeepPartial<DeleteEntryRequest>): DeleteEntryRequest {
-    return DeleteEntryRequest.fromPartial(base ?? {});
+  create(base?: DeepPartial<DeleteEntriesRequest>): DeleteEntriesRequest {
+    return DeleteEntriesRequest.fromPartial(base ?? {});
   },
 
-  fromPartial(object: DeepPartial<DeleteEntryRequest>): DeleteEntryRequest {
-    const message = createBaseDeleteEntryRequest();
+  fromPartial(object: DeepPartial<DeleteEntriesRequest>): DeleteEntriesRequest {
+    const message = createBaseDeleteEntriesRequest();
     message.type = object.type ?? "";
     message.tenantId = object.tenantId ?? "";
     message.id = object.id ?? "";
+    message.filter = (object.filter !== undefined && object.filter !== null)
+      ? EntryFilter.fromPartial(object.filter)
+      : undefined;
     return message;
   },
 };
 
-function createBaseDeleteEntryResponse(): DeleteEntryResponse {
+function createBaseDeleteEntriesResponse(): DeleteEntriesResponse {
   return {};
 }
 
-export const DeleteEntryResponse = {
-  encode(_: DeleteEntryResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
+export const DeleteEntriesResponse = {
+  encode(_: DeleteEntriesResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
     return writer;
   },
 
-  decode(input: _m0.Reader | Uint8Array, length?: number): DeleteEntryResponse {
+  decode(input: _m0.Reader | Uint8Array, length?: number): DeleteEntriesResponse {
     const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
     let end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseDeleteEntryResponse();
+    const message = createBaseDeleteEntriesResponse();
     while (reader.pos < end) {
       const tag = reader.uint32();
       switch (tag >>> 3) {
@@ -107,21 +120,21 @@ export const DeleteEntryResponse = {
     return message;
   },
 
-  fromJSON(_: any): DeleteEntryResponse {
+  fromJSON(_: any): DeleteEntriesResponse {
     return {};
   },
 
-  toJSON(_: DeleteEntryResponse): unknown {
+  toJSON(_: DeleteEntriesResponse): unknown {
     const obj: any = {};
     return obj;
   },
 
-  create(base?: DeepPartial<DeleteEntryResponse>): DeleteEntryResponse {
-    return DeleteEntryResponse.fromPartial(base ?? {});
+  create(base?: DeepPartial<DeleteEntriesResponse>): DeleteEntriesResponse {
+    return DeleteEntriesResponse.fromPartial(base ?? {});
   },
 
-  fromPartial(_: DeepPartial<DeleteEntryResponse>): DeleteEntryResponse {
-    const message = createBaseDeleteEntryResponse();
+  fromPartial(_: DeepPartial<DeleteEntriesResponse>): DeleteEntriesResponse {
+    const message = createBaseDeleteEntriesResponse();
     return message;
   },
 };